requirements.txt
Adafruit-Blinka==8.39.2
adafruit-circuitpython-bitmap-font==2.1.1
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.0
adafruit-circuitpython-requests==4.0.0
adafruit-circuitpython-rgb-display==3.12.4
adafruit-circuitpython-st7735r==1.6.2
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.66.0
Adafruit-PureIO==1.1.11
annotated-types==0.7.0
anyio==4.4.0
arandr==0.1.11
asttokens==2.4.1
attrs==23.2.0
av==12.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
bleak==0.22.2
blinker==1.8.2
certifi==2024.6.2
chardet==5.1.0
charset-normalizer==3.3.2
click==8.1.7
colorzero==2.0
cupshelpers==1.0
dbus-fast==2.21.3
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
docopt==0.6.2
dotenv==0.0.5
executing==2.0.1
fastjsonschema==2.20.0
Flask==3.0.3
gpiozero==2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lgpio==0.2.2.0
libevdev==0.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
meson==1.0.1
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.26.4
olefile==0.46
openai==1.30.5
opencv-python-headless==4.9.0.80
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
picamera==1.13
picamera2==0.3.18
pickleshare==0.7.5
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycups==2.0.1
pydantic==2.7.2
pydantic_core==2.18.3
pyftdi==0.55.4
pygame==2.5.2
Pygments==2.18.0
PyGObject==3.42.2
PyOpenGL==3.1.6
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyserial==3.5
pysmbc==1.0.23
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-prctl==1.8.1
pyudev==0.24.0
pyusb==1.2.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
simplejpeg==1.7.2
six==1.16.0
smbus2==0.4.2
sniffio==1.3.1
soupsieve==2.5
ssh-import-id==5.10
stack-data==0.6.3
sysv-ipc==1.1.0
tinycss2==1.3.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.1
urllib3==2.2.1
v4l2-python3==0.3.4
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
wraptext==0.1.0b1
yarg==0.1.9
